+++ NEWS.md
@@ -28,6 +28,7 @@
     - `plot_graph()` => `vis_drake_graph()`
     - `read_config()` => `read_drake_config()`
     - `read_graph()` => `read_drake_graph()`
+    - `read_plan()` => `read_drake_plan()`
     - `render_graph()` => `render_drake_graph()`
     - `session()` => `drake_session()`
     - `summaries()` => `plan_summaries()`